refactor: 기사 필터링 기능 구현 위치 수정
픽업 스케줄에 따라 가능한 기사 목록을 필터링하는 기능이 기존의 repository 계층에서 처리되던 것을 service 계층에서 처리되도록 변경하였습니다.
