+++ BE/iDrop/src/main/java/ifive/idrop/repository/DriverRepository.java
@@ -30,23 +30,6 @@ public class DriverRepository {
                 .getResultList();
     }
 
-    public List<Driver> findDriversBySchedule(DriverListRequest driverListRequest) {
-        RequestSchedule requestSchedule = ScheduleUtils.parseToList(driverListRequest.getSchedule());
-
-        List<Driver> availableDrivers = new ArrayList<>();
-        List<Driver> drivers = findAllDrivers();
-        for (Driver driver : drivers) {
-            List<PickUp> pickUpList = pickUpRepository.findReservedPickUpsByDriver(driver.getId());
-            List<LocalDateTime> reservedSchedule = pickUpList.stream()
-                    .map(PickUp::getReservedTime)
-                    .toList();
-            List<WorkHours> workHoursList = driver.getWorkHoursList();
-            if (requestSchedule.isAvailable(workHoursList, reservedSchedule)) {
-                availableDrivers.add(driver);
-            }
-        }
-        return availableDrivers;
-    }
     public List<PickUpInfo> findRunningPickInfo(Long driverId) {
         String query = "SELECT pui\n" +
                 "FROM PickUpInfo pui\n" +
