+++ BE/iDrop/src/main/java/ifive/idrop/service/DriverService.java
@@ -1,20 +1,28 @@
 package ifive.idrop.service;
 
+import ifive.idrop.annotation.Login;
 import ifive.idrop.dto.response.BaseResponse;
 import ifive.idrop.dto.request.DriverInformation;
 import ifive.idrop.dto.request.DriverListRequest;
 import ifive.idrop.dto.response.DriverDetailResponse;
 import ifive.idrop.entity.Driver;
+import ifive.idrop.entity.PickUp;
+import ifive.idrop.entity.WorkHours;
 import ifive.idrop.exception.CommonException;
 import ifive.idrop.exception.ErrorCode;
 import ifive.idrop.repository.DriverRepository;
+import ifive.idrop.repository.PickUpRepository;
+import ifive.idrop.util.RequestSchedule;
+import ifive.idrop.util.ScheduleUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import ifive.idrop.dto.CurrentPickUpResponse;
+import ifive.idrop.dto.response.CurrentPickUpResponse;
 import ifive.idrop.entity.PickUpInfo;
 
+import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 
 
@@ -23,10 +31,25 @@ import java.util.List;
 @Service
 public class DriverService {
     private final DriverRepository driverRepository;
+    private final PickUpRepository pickUpRepository;
 
     @Transactional(readOnly = true)
     public List<Driver> searchAvailableDrivers(DriverListRequest driverListRequest) {
-        return driverRepository.findDriversBySchedule(driverListRequest);
+        RequestSchedule requestSchedule = ScheduleUtils.parseToList(driverListRequest.getSchedule());
+
+        List<Driver> availableDrivers = new ArrayList<>();
+        List<Driver> drivers = driverRepository.findAllDrivers();
+        for (Driver driver : drivers) {
+            List<PickUp> pickUpList = pickUpRepository.findReservedPickUpsByDriver(driver.getId());
+            List<LocalDateTime> reservedSchedule = pickUpList.stream()
+                    .map(PickUp::getReservedTime)
+                    .toList();
+            List<WorkHours> workHoursList = driver.getWorkHoursList();
+            if (requestSchedule.isAvailable(workHoursList, reservedSchedule)) {
+                availableDrivers.add(driver);
+            }
+        }
+        return availableDrivers;
     }
 
     @Transactional
@@ -50,6 +73,10 @@ public class DriverService {
                 runningPickInfo.stream()
                         .map(CurrentPickUpResponse::of)
                         .toList());
+    }
+
+    @Transactional(readOnly = true)
+    public void subscribeList(Long driverId) {
     }
 
 }
