+++ BE/iDrop/src/main/java/ifive/idrop/service/ParentService.java
@@ -3,7 +3,7 @@ package ifive.idrop.service;
 
 import ifive.idrop.dto.response.BaseResponse;
 import ifive.idrop.dto.request.SubscribeRequest;
-import ifive.idrop.dto.CurrentPickUpResponse;
+import ifive.idrop.dto.response.CurrentPickUpResponse;
 
 import ifive.idrop.entity.*;
 import ifive.idrop.entity.enums.PickUpStatus;
